Improved based on review 1.
