+++ python/sglang/srt/layers/quantization/modelopt/schemes/__init__.py
@@ -1,11 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
-# FP4 first (shared MoE symbols for FP8); then FP8; then mixed precision.
 from sglang.srt.layers.quantization.modelopt.schemes.modelopt_fp4 import (
-    ACT_STR_TO_TYPE_MAP,
     ACTIVATION_SCHEMES,
     MOE_NVFP4_DISPATCH,
-    ActivationType,
     ModelOptFp4Config,
     ModelOptFp4LinearMethod,
     ModelOptNvFp4FusedMoEMethod,
@@ -21,6 +18,10 @@ from sglang.srt.layers.quantization.modelopt.schemes.modelopt_fp8 import (
 )
 from sglang.srt.layers.quantization.modelopt.schemes.modelopt_mixed_precision import (
     ModelOptMixedPrecisionConfig,
+)
+from sglang.srt.layers.quantization.modelopt.utils import (
+    ACT_STR_TO_TYPE_MAP,
+    ActivationType,
 )
 
 __all__ = [
